requirements.txt
Flask==3.1.0
flask-cors==5.0.1
pandas==2.2.3
psycopg2-binary==2.9.10
pytest==8.3.5
python-dotenv==1.1.0
selenium==4.30.0
requests==2.32.3
pdfplumber==0.11.5